[CT-52] 도메인 모델 객체 생성 돕는 메서드 필드 모아놓은 클래스 분리, 테스트 설명 주석 추가
